Refs #4776. Removing RemoteWriterAttributes.

+++ include/fastrtps/subscriber/SampleInfo.h
@@ -43,7 +43,7 @@ public:
 	//!Sample kind.
 	rtps::ChangeKind_t sampleKind;
 	//!Ownership Strength of the writer of the sample (0 if the ownership kind is set to SHARED_OWNERSHIP_QOS).
-	uint16_t ownershipStrength;
+	uint32_t ownershipStrength;
 	//!Source timestamp of the sample.
 	rtps::Time_t sourceTimestamp;
 	//!InstanceHandle of the data
